Remove Api from Core.Test
