+++ test/Common/Common.csproj
@@ -14,10 +14,9 @@
     <PackageReference Include="AutoFixture.Xunit2" Version="$(AutoFixtureXUnit2Version)" />
     <PackageReference Include="AutoFixture.AutoNSubstitute" Version="$(AutoFixtureAutoNSubstituteVersion)" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="$(MicrosoftNetTestSdkVersion)" />
-    <PackageReference Include="Kralizek.AutoFixture.Extensions.MockHttp" Version="1.2.0"/>
+    <PackageReference Include="Kralizek.AutoFixture.Extensions.MockHttp" Version="1.2.0" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\..\src\Core\Core.csproj"/>
-    <ProjectReference Include="..\..\src\Api\Api.csproj"/>
+    <ProjectReference Include="..\..\src\Core\Core.csproj" />
   </ItemGroup>
 </Project>
