--- test/Common/Helpers/Factories.cs
@@ -1,16 +1,0 @@
-﻿using Bit.Core.Settings;
-using Microsoft.Extensions.Configuration;
-
-namespace Bit.Test.Common.Helpers.Factories
-{
-    public static class GlobalSettingsFactory
-    {
-        public static GlobalSettings GlobalSettings { get; } = new GlobalSettings();
-        static GlobalSettingsFactory()
-        {
-            var configBuilder = new ConfigurationBuilder().AddUserSecrets<Bit.Api.Startup>();
-            var Configuration = configBuilder.Build();
-            ConfigurationBinder.Bind(Configuration.GetSection("GlobalSettings"), GlobalSettings);
-        }
-    }
-}
