+++ test/Core.Test/Helpers/Factories.cs
@@ -1,6 +1,7 @@
-﻿using Bit.Infrastructure.EntityFramework.Repositories;
-using Bit.Test.Common.Helpers.Factories;
+﻿using Bit.Core.Settings;
+using Bit.Infrastructure.EntityFramework.Repositories;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Configuration;
 
 namespace Bit.Core.Test.Helpers.Factories
 {
@@ -21,6 +22,17 @@ namespace Bit.Core.Test.Helpers.Factories
                 var mySqlConnectionString = globalSettings.MySql.ConnectionString;
                 Options.Add(new DbContextOptionsBuilder<DatabaseContext>().UseMySql(mySqlConnectionString, ServerVersion.AutoDetect(mySqlConnectionString)).Options);
             }
+        }
+    }
+
+    public static class GlobalSettingsFactory
+    {
+        public static GlobalSettings GlobalSettings { get; } = new GlobalSettings();
+        static GlobalSettingsFactory()
+        {
+            var configBuilder = new ConfigurationBuilder().AddUserSecrets<Bit.Api.Startup>();
+            var Configuration = configBuilder.Build();
+            ConfigurationBinder.Bind(Configuration.GetSection("GlobalSettings"), GlobalSettings);
         }
     }
 }
